First revision of axes, swords and loot ball's
The axes still need some work at this point but they are now in a
working state. Also added loot balls to go with the new wip harvesting
system for the axes and soon the mining tools. And the swords now work
as well. Could also still use some work.